requirements.txt
asgiref==3.8.1
astroid==3.3.9
blinker==1.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
dill==0.4.0
Django==5.2
django-ckeditor==6.7.2
django-js-asset==3.1.2
Flask==3.1.0
flask-cors==5.0.1
idna==3.10
isort==6.0.1
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
mccabe==0.7.0
pillow==11.2.1
platformdirs==4.3.7
pylint==3.3.6
pylint-celery==0.3
pylint-django==2.6.1
pylint-plugin-utils==0.8.2
python-dateutil==2.9.0.post0
requests==2.32.3
six==1.17.0
sqlparse==0.5.3
tomlkit==0.13.2
tzdata==2025.2
urllib3==2.4.0
Werkzeug==3.1.3
